projekt_mappe/requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
pyyaml
openpyxl
pytest

